+++ lib/async_extension.dart
@@ -4,3 +4,4 @@ library async_extension;
 export 'dart:async';
 
 export 'src/async_extension_base.dart';
+export 'src/async_extension_try.dart';